# flood-alert-java/src/extra/api.yaml
swagger: '2.0'
info:
  title: Flood Alert
  description: Flood Alert Services
  version: 0.1.0
schemes:
  - http
basePath: /flood-alert/restApi
produces:
  - application/json
paths:
  /version:
    get:
      summary: Returns Api version
      tags:
        - Misc
      responses:
        '200':
          description: Current version
          schema:
            $ref: '#/definitions/VersionResponse'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    x-swagger-router-controller: version
  /alerts:
    get:
      description: |
        Gets `Alerts` objects.
      parameters:
        - $ref: '#/parameters/Page'
        - $ref: '#/parameters/PageSize'
        - $ref: '#/parameters/Lat'
        - $ref: '#/parameters/Lon'
        - $ref: '#/parameters/Date'
      tags:
        - Alerts
      responses:
        '200':
          description: Successful response
          schema:
            $ref: '#/definitions/Alerts'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    post:
      consumes:
        - application/json
      description: |
        Create a new alert.
      tags:
        - Alerts
      parameters:
        - $ref: '#/parameters/Alert'
      responses:
        '201':
          description: Successful response
          schema:
            $ref: '#/definitions/Alert'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  /schema:
    x-swagger-pipe: swagger_raw
definitions:
  Alerts:
    type: object
    properties:
      count:
        type: integer
      alerts:
        type: array
        items:
          $ref: '#/definitions/Alert'
  Alert:
    type: object
    properties:
      id:
        type: integer
        format: int64
      date:
        type: string
        format: date
      message:
        type: string
      severity:
        type: string
        enum:
          - WARNING
          - CAUTION
          - DANGER
      latitude:
        type: number
        format: double
      longitude:
        type: number
        format: double
  VersionResponse:
    type: object
    properties:
      version:
        type: string
  Error:
    type: object
    properties:
      message:
        type: string
parameters:
  Alert:
    name: Alert
    in: body
    schema:
      $ref: '#/definitions/Alert'
  Page:
    name: page
    in: query
    description: Page number (1 based).
    type: integer
    default: 1
    minimum: 1
  PageSize:
    name: pageSize
    in: query
    type: integer
    default: 10
    maximum: 200
  Lat:
    name: lat
    in: query
    description: Latitude
    required: false
    type: number
    format: double
  Lon:
    name: lon
    in: query
    description: Longitude
    required: false
    type: number
    format: double
  Date:
    name: date
    in: query
    description: Date
    required: false
    type: string
    format: date